+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 1.31.2 - 2024-10-13
+* Added button to jump to HEAD. Thanks [@QrackEE](https://github.com/QrackEE).
 * Add collapse/expand buttons to commit diff view. Thanks [@LariWa](https://github.com/LariWa)
 * Added sticky header option. Thanks [@bendera](https://github.com/bendera)